requirements.txt
Django==4.2.3
djangorestframework==3.14.0
matplotlib==3.8.1
numpy==1.25.0
Pillow==10.0.1
Pillow==10.1.0
pydicom==2.3.1
